x/maki/types/genesis_test.go
package types_test

import (
	"testing"

	"github.com/mkXultra/maki_chain/x/maki/types"
	"github.com/stretchr/testify/require"
)

func TestGenesisState_Validate(t *testing.T) {
	for _, tc := range []struct {
		desc     string
		genState *types.GenesisState
		valid    bool
	}{
		{
			desc:     "default is valid",
			genState: types.DefaultGenesis(),
			valid:    true,
		},
		{
			desc: "valid genesis state",
			genState: &types.GenesisState{

				MakiList: []types.Maki{
					{
						Id: 0,
					},
					{
						Id: 1,
					},
				},
				MakiCount: 2,
				// this line is used by starport scaffolding # types/genesis/validField
			},
			valid: true,
		},
		{
			desc: "duplicated maki",
			genState: &types.GenesisState{
				MakiList: []types.Maki{
					{
						Id: 0,
					},
					{
						Id: 0,
					},
				},
			},
			valid: false,
		},
		{
			desc: "invalid maki count",
			genState: &types.GenesisState{
				MakiList: []types.Maki{
					{
						Id: 1,
					},
				},
				MakiCount: 0,
			},
			valid: false,
		},
		// this line is used by starport scaffolding # types/genesis/testcase
	} {
		t.Run(tc.desc, func(t *testing.T) {
			err := tc.genState.Validate()
			if tc.valid {
				require.NoError(t, err)
			} else {
				require.Error(t, err)
			}
		})
	}
}
